--- src/components/centralWhite.css
@@ -1,12 +1,0 @@
-
-
-/* CONTENEDOR DE BOTONES DESAYUNO Y ALMUERZO */
-.containerButtonsMenu{
-    width: 100%;
-    justify-content: space-around;
-    justify-content: center;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
