+++ src/secctions/breakfast.js
@@ -1,40 +1,37 @@
 import React, { Component } from 'react';
-import MenuCards from './menuCards.js'
+import MenuCards from '../components/menuCards.js'
 import Data from "../data/menu.json";
 //import db from './confg/firebase';
 
 
-
-
-class LunchMenu extends Component {
+class Breackfast extends Component {
   constructor() {
     super() ;
-    this.state = {menu: Data.Lunch} 
+    this.state = {menu: Data.Breakfast} 
     console.log()
   }
 
-
-   /* constructor() {
-      super();
+    /*constructor(props) {
+      super(props);
       this.state = {
         menuA: []
       }
     }
     
-    componentDidMount(){
-      db.collection('Menualmuerzo').get().then((querySnapshot) => {
+    /*componentDidMount(){
+      db.collection('MenuDesayuno').get().then((querySnapshot) => {
         const data = querySnapshot.docs.map(doc => doc.data());
         this.setState({
           menuA:data
         })
       
-        console.log("STATE:", this.state);
+        console.log("STATE:", this.state.menuA);
       })
           
     }*/
-    render(){
-      return (      
-        this.state.menu.map((e =>
+    render(){         
+      return (
+          this.state.menu.map((e =>
           <MenuCards
           key = {e.id}
           img = {e.img}
@@ -43,8 +40,7 @@ class LunchMenu extends Component {
           />)
           )
       )
-   }    
-      
+   }        
     }
   
-    export default LunchMenu;
+export default Breackfast
