+++ src/secctions/lunch.js
@@ -1,37 +1,40 @@
 import React, { Component } from 'react';
-import MenuCards from './menuCards.js'
+import MenuCards from '../components/menuCards.js'
 import Data from "../data/menu.json";
 //import db from './confg/firebase';
 
 
-class Breackfast extends Component {
+
+
+class LunchMenu extends Component {
   constructor() {
     super() ;
-    this.state = {menu: Data.Breakfast} 
+    this.state = {menu: Data.Lunch} 
     console.log()
   }
 
-    /*constructor(props) {
-      super(props);
+
+   /* constructor() {
+      super();
       this.state = {
         menuA: []
       }
     }
     
-    /*componentDidMount(){
-      db.collection('MenuDesayuno').get().then((querySnapshot) => {
+    componentDidMount(){
+      db.collection('Menualmuerzo').get().then((querySnapshot) => {
         const data = querySnapshot.docs.map(doc => doc.data());
         this.setState({
           menuA:data
         })
       
-        console.log("STATE:", this.state.menuA);
+        console.log("STATE:", this.state);
       })
           
     }*/
-    render(){         
-      return (
-          this.state.menu.map((e =>
+    render(){
+      return (      
+        this.state.menu.map((e =>
           <MenuCards
           key = {e.id}
           img = {e.img}
@@ -40,7 +43,8 @@ class Breackfast extends Component {
           />)
           )
       )
-   }        
+   }    
+      
     }
   
-export default Breackfast
+    export default LunchMenu;
